# docs/swagger.yaml
definitions:
...
  personalfinancedss_internal_module_cashflow_account_dto.AccountResponse:
    properties:
      account_name:
        type: string
      account_number_masked:
        type: string
      account_type:
        type: string
      available_balance:
        type: number
      created_at:
        type: string
      currency:
        type: string
      current_balance:
        type: number
      id:
        type: string
      include_in_net_worth:
        type: boolean
      institution_name:
        type: string
      is_active:
        type: boolean
      is_primary:
        type: boolean
      last_synced_at:
        type: string
      sync_error_message:
        type: string
      sync_status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  personalfinancedss_internal_module_cashflow_account_dto.AccountsListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/personalfinancedss_internal_module_cashflow_account_dto.AccountResponse'
        type: array
      total:
        type: integer
    type: object
  personalfinancedss_internal_module_cashflow_account_dto.CreateAccountRequest:
    properties:
      account_name:
        maxLength: 255
        minLength: 1
        type: string
      account_number_masked:
        maxLength: 50
        type: string
      account_type:
        enum:
        - cash
        - bank
        - savings
        - credit_card
        - investment
        - crypto_wallet
        type: string
      available_balance:
        type: number
      currency:
        type: string
      current_balance:
        type: number
      include_in_net_worth:
        type: boolean
      institution_name:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      is_primary:
        type: boolean
    required:
    - account_name
    - account_type
    type: object
  personalfinancedss_internal_module_cashflow_account_dto.UpdateAccountRequest:
    properties:
      account_name:
        maxLength: 255
        minLength: 1
        type: string
      account_number_masked:
        maxLength: 50
        type: string
      account_type:
        enum:
        - cash
        - bank
        - savings
        - credit_card
        - investment
        - crypto_wallet
        type: string
      available_balance:
        type: number
      currency:
        type: string
      current_balance:
        type: number
      include_in_net_worth:
        type: boolean
      institution_name:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      is_primary:
        type: boolean
      sync_error_message:
        type: string
      sync_status:
        enum:
        - active
        - error
        - disconnected
        type: string
    type: object
...
  personalfinancedss_internal_shared.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      error:
        type: string
      message:
        type: string
    type: object
...
info:
  contact:
    email: support@personalfinancedss.com
    name: API Support
    url: http://www.swagger.io/support
  description: Personal Finance Decision Support System API for managing accounts,
    transactions, budgets, goals, and investments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Personal Finance DSS API
  version: "1.0"
paths:
  /api/v1/accounts:
    post:
      consumes:
      - application/json
      description: Create a new financial account for the authenticated user
      parameters:
      - description: Account data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/personalfinancedss_internal_module_cashflow_account_dto.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_account_dto.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new account
      tags:
      - accounts
  /api/v1/accounts/:
    get:
      consumes:
      - application/json
      description: Get all accounts for the authenticated user with optional filters
      parameters:
      - description: Filter by account type
        enum:
        - cash
        - bank
        - savings
        - credit_card
        - investment
        - crypto_wallet
        in: query
        name: account_type
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      - description: Filter by primary status
        in: query
        name: is_primary
        type: boolean
      - description: Include deleted accounts
        in: query
        name: include_deleted
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_account_dto.AccountsListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my accounts
      tags:
      - accounts
  /api/v1/accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete an account for the authenticated user
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an account
      tags:
      - accounts
    get:
      consumes:
      - application/json
      description: Get a specific account by ID (must belong to authenticated user)
      parameters:
      - description: Account ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_account_dto.AccountResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get account by ID
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: Update an account for the authenticated user
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/personalfinancedss_internal_module_cashflow_account_dto.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_account_dto.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an account
      tags:
      - accounts
...
securityDefinitions:
  BearerAuth:
    description: JWT Authorization header using the Bearer scheme. Enter 'Bearer'
      [space] and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"